pyproject.toml
[project]
name = "real-estate-agent"
version = "0.1.0"
description = "A simple chat agent using AutoGen for real estate assistance"
authors = [
    {name = "toreleon",email = "levietthang0512@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0", 
    "python-dotenv>=1.0.0", 
    "autogen-agentchat>=0.7.1,<0.8.0", 
    "autogen-ext[openai]>=0.7.1,<0.8.0", 
    "motor (>=3.7.1,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
